navigation-timing/nav2_test_attributes_values.html

<!DOCTYPE html>
<html>
    <head>
        <meta charset="utf-8">
        <title>window.performance.getEntriesByType("navigation") attributes</title>
        <link rel="author" title="Google" href="http://www.google.com/" />
        <link rel="help" href="http://www.w3.org/TR/navigation-timing-2/#sec-PerformanceNavigationTiming"/>
        <script src="/resources/testharness.js"></script>
        <script src="/resources/testharnessreport.js"></script>
    </head>
    <body>
        <h1>Description</h1>
        <p>This test validates that some of nav timing 2 instance's timing-related attributes' values are in certain order and some are of expected values.</p>

        <div id="log"></div>

        <script>
        var navTimingEventOrder = [
            'startTime',
            'redirectStart',
            'redirectEnd',
            'fetchStart',
            'domainLookupStart',
            'domainLookupEnd',
            'connectStart',
            'connectEnd',
            'requestStart',
            'responseStart',
            'responseEnd',
            'domInteractive',
            'domContentLoadedEventStart',
            'domContentLoadedEventEnd',
            'domComplete',
            'loadEventStart',
            'loadEventEnd'
        ];
        async_test(function (t) {
            var observer = new PerformanceObserver(
                t.step_func(function (entryList) {
                    var entries = entryList.getEntries();
                    assert_equals(entries[0].entryType, "navigation",
                        "entryType expected to be: navigation");
                    assert_equals(entries[0].name, "document",
                        "name expected to be: document");
                    assert_equals(entries[0].startTime, 0,
                        "startTime expected to be: 0");
                    assert_equals(entries[0].duration, entries[0].loadEventEnd,
                        "expected duration to be equal to loadEventEnd");

                    for (var i = 0; i < navTimingEventOrder.length - 1; i++) {
                        assert_true(entries[0][navTimingEventOrder[i]] <= entries[0][navTimingEventOrder[i + 1]],
                            "expected " + navTimingEventOrder[i] + " to be no greater than " + navTimingEventOrder[i + 1] + ".");
                    }
                    observer.disconnect();
                    t.done();
                })
            );
            observer.observe({entryTypes: ["navigation"]});

        }, "Performance navigation timing instance's value is reasonable.");
        </script>
    </body>
</html>
